Read is done with JSON.NET serializer.
As far as exercise says "REST service endpoint to return a user object", it states that 'CREATE' and 'UPDATE' and 'DELETE' operations are not needed.

It is also extra stressed out by 'it doesn't need to fetch from a database or anything. Just pretend that your service is fetching a user and returning it.', to CREATE, UPDATE and DELETE you need some persistent storage.

Not sure why "for incoming requests and responses." as far as requests are all in query string with GET.